feat: Implement user order history and profile management components
- Added HistorialPedidos component to display user's order history with detailed statistics and error handling.
- Created InformacionUsuario component to show user profile information and basic statistics.
- Developed PerfilUsuario component for editing user profile details and changing password with validation.
- Integrated API calls for fetching user orders and updating user data, including profile picture uploads.
- Enhanced user experience with loading states, error messages, and success notifications.
